Moved username from Name to Internet, which seems like a slightly better spot.
